If profile photo is deleted it is updated on profile screen
If user name is clicked on top right it goes to Profile page for the user
Users are not listed until user creates a profile
Photo upload is now easier
Users can change password on Profile page
Sexist fields are hidden from users for now.
Change password is a Partial view for now.